TGexplorer_bot/src/handlers/media.py
from telegram import Update
from telegram.ext import ContextTypes

async def search_media(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    keyword = ' '.join(context.args)
    if not keyword:
        await update.message.reply_text("Please provide a keyword to search for.")
        return

    # Here you would implement the logic to search for multimedia messages
    # containing the keyword in their captions or associated text.
    # This is a placeholder for the actual search logic.
    
    results = []  # This should be populated with the search results

    if results:
        response = "\n".join(results)
        await update.message.reply_text(f"Found the following multimedia messages:\n{response}")
    else:
        await update.message.reply_text("No multimedia messages found with that keyword.")