TGexplorer_bot/requirements.txt
aiohttp
python-telegram-bot==20.0
beautifulsoup4
requests
pandas